requirements.txt
Django
gunicorn
psycopg2
dj-database-url
python-dotenv
requests
